refactor modules/auth
