+++ src/modules/auth/controllers/auth.controller.ts
@@ -1,13 +1,14 @@
 import { Controller, Get, Post, Req, Res } from '@nestjs/common'
 import { Response } from 'express'
-import { ApiOAuth2, ApiResponse, ApiTags } from '@nestjs/swagger'
+import { ApiBearerAuth, ApiResponse, ApiTags } from '@nestjs/swagger'
 import { AuthService } from '../services/auth.service.js'
-import { Public } from '../../permissions/permissions.decorator.js'
+import { Public } from '../../permissions/decorators/permissions.decorator.js'
 import { Request } from '../guards/auth.guard.js'
 import { AuthTransformer } from '../transformers/auth.transformer.js'
-import { UserTransformer, UserTransformerType } from '../../users/transformers/user.transformer.js'
+import { UserTransformer, type UserTransformerType } from '../../users/transformers/user.transformer.js'
+import { getUserInfoResponse } from '../docs/auth-response.docs.js'
 
-@ApiTags('Auth')
+@ApiTags('Authentication')
 @Controller({
   path: 'auth',
   version: ''
@@ -19,7 +20,10 @@ export class AuthController {
 
   @Post('/token')
   @Public()
-  public async createToken (@Req() req: Request, @Res() res: Response): Promise<void> {
+  public async createToken (
+    @Req() req: Request,
+    @Res() res: Response
+  ): Promise<void> {
     try {
       const token = await this.authService.signIn(req, res)
 
@@ -33,13 +37,11 @@ export class AuthController {
   }
 
   @Get('/userinfo')
-  @ApiOAuth2([])
-  @ApiResponse({
-    status: 200,
-    description: 'The user info has been successfully retrieved.',
-    type: UserTransformerType
-  })
-  public async getUserInfo (@Req() req: Request): Promise<UserTransformerType> {
+  @ApiBearerAuth()
+  @ApiResponse(getUserInfoResponse)
+  public async getUserInfo (
+    @Req() req: Request
+  ): Promise<UserTransformerType> {
     const user = await this.authService.getUserInfo(req)
     return new UserTransformer().item(user)
   }
